remove icon from fahrspuren
